Add bumerang docker image and smoke test
Created a Centos7 Python 3.5 docker image. Centos7 was choosen because it is
smaller than ubuntu server and hence it may be nice for clustering. The smoke
test was is given with the /health end point. A make file was added to allow
easy access to create and test the server, however, more work should be done to
create an easy cleanup for the left over containers and images.