Fix empty result from query select
When the query selected returned an empty list, the assert that assumed the
result had to be exacly one failed for a given id. The logic here is there for
a given unique id, less or more then one results should be returned, but only
one was accounted for here. Added error handling as opposed to an assert so the
endpoint returns an error message instead of crashing.